+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 # Unreleased
 
+- Adds `utils.file_hash`
 - Adds `Execution().deferred()` to prevent automatic dispatch
 - Respect CLI context order